checkAuth middleware created
