+++ backend/app.js
@@ -4,21 +4,35 @@ const path = require("path");
 const cookieParser = require("cookie-parser");
 const logger = require("morgan");
 const cors = require("cors")
-const admin = require("firebase-admin")
+// const admin = require("firebase-admin")
 
 const indexRouter = require("./routes/index");
 const usersRouter = require("./routes/users");
 const receiptsRouter = require("./routes/receipts");
 const authRouter = require("./routes/auth");
 
+// const serviceAccount = require("./config/fbServiceAccountKey.json");
 const app = express();
 
-const serviceAccount = require("./config/fbServiceAccountKey.json");
+// admin.initializeApp({
+//   credential: admin.credential.cert(serviceAccount),
+//   databaseURL: "https://capstone-pantry.firebaseio.com"
+// });
 
-admin.initializeApp({
-  credential: admin.credential.cert(serviceAccount),
-  databaseURL: "https://capstone-pantry.firebaseio.com"
-});
+// const checkAuth = (req, res, next) => {
+//   if (req.headers.authtoken) {
+//     admin.auth().verifyIdToken(req.headers.authtoken)
+//       .then((eto) => {
+//         console.log(eto)
+//         next()
+//       }).catch(() => {
+//         res.status(403).send('Unauthorized')
+//       });
+//   } else {
+//     res.status(403).send('Unauthorized!')
+//     return
+//   }
+// }
 
 
 app.use(cors())
@@ -28,6 +42,8 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
 
+
+// app.use('/', checkAuth)
 app.use("/", indexRouter);
 app.use("/users", usersRouter);
 app.use("/receipts", receiptsRouter);
